requirements.txt
fastapi
uvicorn
boto3
mysql-connector-python
python-dotenv